+++ index.js
@@ -17,11 +17,12 @@ app.use(cookieParser())
 app.use('/', require(path.join(__dirname, "/routes/pages/root.js")));
 app.use('/api/components', require(path.join(__dirname, "/routes/api/components.js")));
 app.use('/api/auth', require(path.join(__dirname, "/routes/api/authentication.js")));
-app.use('/api/cars', require(path.join(__dirname, "/routes/api/create-cars-data.js")));
+app.use('/api/cars', require(path.join(__dirname, "/routes/api/cars.js")));
 app.use('/api/subscribe', require(path.join(__dirname, "/routes/api/email-subscribe.js")));
 app.use('/api/account', require(path.join(__dirname, "/routes/api/account.js")));
 app.use('/api/booking', require(path.join(__dirname, "/routes/api/booking.js")));
 app.use('/api/invoice', require(path.join(__dirname, "/routes/api/invoice.js")));
+app.use('/api/transaction', require(path.join(__dirname, "/routes/api/transaction.js")));
 
 
 app.listen(port, () => {