+++ routes/api/booking.js
@@ -2,7 +2,6 @@ const express = require('express');
 const router = express.Router();
 const { authCheck } = require('../../middleware/authentication/authentication');
 const { orderCreate } = require('../../models/functions/booking/order-create.js')
-const { transactionComplete } = require('../../models/functions/booking/transaction-complete.js')
 const { orderComplete } = require('../../models/functions/booking/order-complete.js')
 const { carBookDates } = require('../../models/functions/booking/book-car-dates.js')
 const { orderCancel } = require('../../models/functions/booking/cancel-order.js')
@@ -12,9 +11,6 @@ router.post('/order-create', authCheck, orderCreate);
 
 // ! Order Complete API Endpoint
 router.post('/order-complete', authCheck, orderComplete);
-
-// ! Order Payment Complete API Endpoint
-router.post('/transaction-complete', authCheck, transactionComplete)
 
 // ! Book Car Dates after payment completion
 router.post('/book-car-dates', authCheck, carBookDates)